contour canvas mode prep refactor
