+++ examples/contour-extractor/index.ts
@@ -1,5 +1,5 @@
 import TextRenderer from '../../src/index'
-
+import TextPreviewCanvas2D from './TextPreviewCanvas2D'
 import BarlowBold from '../fonts/Barlow-Bold.ttf'
 //import fontPath from '../fonts/FiraCode-Bold.otf'
 //import fontPath from '../fonts/ArefRuqaa-Bold.ttf'
@@ -7,13 +7,14 @@ import BarlowBold from '../fonts/Barlow-Bold.ttf'
 import ScheherazadeBold from '../fonts/Scheherazade-Bold.ttf'
 import AmiriBold from '../fonts/Amiri-Bold.ttf'
 
-import { Path, BoundingBox } from 'opentype.js'
+import { Path } from 'opentype.js'
+import { TextDirection, TextAlign } from '../../src/TextOptions'
 
 import Ruler, { RulerDirection } from './Ruler'
 import TextEditor from '../common/TextEditor'
 
 const canvas = document.querySelector('canvas#viewport') as HTMLCanvasElement
-const context = canvas.getContext('2d')!
+const textPrevCanvas = new TextPreviewCanvas2D(canvas)
 const rulerHorizontal = new Ruler(
   'canvas.ruler.ruler-horizontal',
   RulerDirection.Horizontal
@@ -75,11 +76,7 @@ async function main() {
     const width = window.innerWidth - 30
     const height = window.innerHeight - 30
     const pixelRatio = window.devicePixelRatio
-
-    canvas.width = width * pixelRatio
-    canvas.height = height * pixelRatio
-
-    context.setTransform(pixelRatio, 0, 0, pixelRatio, 0, 0)
+    textPrevCanvas.resize(width, height, pixelRatio)
 
     render()
   }
@@ -93,62 +90,7 @@ async function main() {
 
     rulerHorizontal.render(maxWidth)
     rulerVertical.render(maxHeight)
-
-    context.restore()
-    context.save()
-
-    // Clear
-    context.fillStyle = '#eee'
-    context.fillRect(0, 0, canvas.width, canvas.height)
-
-    // Draw container
-    context.fillStyle = '#fff'
-    context.fillRect(0, 0, maxWidth, maxHeight)
-
-    // Draw lines
-    context.strokeStyle = '#eee'
-
-    for (let x = 0; x < maxWidth; x += 50) {
-      context.beginPath()
-      context.moveTo(x, 0)
-      context.lineTo(x, canvas.height)
-      context.stroke()
-      context.closePath()
-    }
-
-    for (let y = 0; y < maxHeight; y += 50) {
-      context.beginPath()
-      context.moveTo(0, y)
-      context.lineTo(canvas.width, y)
-      context.stroke()
-      context.closePath()
-    }
-
-    lines.forEach(paths => {
-      const boundingBoxes = paths.map(path => path.getBoundingBox())
-      const mergedPath = paths.reduce<Path>((acc, path) => {
-        acc.extend(path)
-        return acc
-      }, new Path())
-
-      renderLine(mergedPath, boundingBoxes, lineHeight)
-    })
-  }
-
-  function renderLine(
-    path: Path,
-    boundingBoxes: BoundingBox[],
-    lineHeight: number
-  ) {
-    context.translate(0, lineHeight)
-
-    path.draw(context)
-
-    context.strokeStyle = 'rgba(0, 0, 255, 1)'
-    // Draw bounding boxes
-    boundingBoxes.forEach(bb => {
-      context.strokeRect(bb.x1, bb.y1, bb.x2 - bb.x1, bb.y2 - bb.y1)
-    })
+    textPrevCanvas.draw(lines, maxWidth, maxHeight, lineHeight)
   }
 
   update()
